fix: plugin name error global search and replace
