+++ internal/pkg/plugin/reposcaffolding/reposcaffolding.go
@@ -1,0 +1,1 @@
+package reposcaffolding
