--- internal/pkg/plugin/scaffolding/scaffolding.go
@@ -1,1 +1,0 @@
-package scaffolding
