+++ internal/pkg/show/config/embed_gen.go
@@ -66,8 +66,8 @@ var (
 	//go:embed plugins/openldap.yaml
 	OpenldapDefaultConfig string
 
-	//go:embed plugins/scaffolding.yaml
-	ScaffoldingDefaultConfig string
+	//go:embed plugins/repo-scaffolding.yaml
+	RepoScaffoldingDefaultConfig string
 
 	//go:embed plugins/tekton.yaml
 	TektonDefaultConfig string
@@ -102,7 +102,7 @@ var pluginDefaultConfigs = map[string]string{
 	"jira-github-integ":              JiraGithubIntegDefaultConfig,
 	"kube-prometheus":                KubePrometheusDefaultConfig,
 	"openldap":                       OpenldapDefaultConfig,
-	"scaffolding":                    ScaffoldingDefaultConfig,
+	"repo-scaffolding":               RepoScaffoldingDefaultConfig,
 	"tekton":                         TektonDefaultConfig,
 	"trello-github-integ":            TrelloGithubIntegDefaultConfig,
 	"trello":                         TrelloDefaultConfig,
