+++ internal/pkg/show/config/plugins/repo-scaffolding.yaml
@@ -1,6 +1,6 @@
 tools:
 # name of the tool
-- name: scaffolding
+- name: reposcaffolding
   # id of the tool instance
   instanceID: default
   # format: name.instanceID; If specified, dtm will make sure the dependency is applied first before handling this tool.
